requirements.txt
playwright==1.42.0
beautifulsoup4==4.12.3
html2text==2024.2.26
tenacity==8.2.3
langchain-text-splitters==0.3.6
tiktoken==0.8.0
sentence-transformers==3.3.1
pinecone-client==3.2.2
python-dotenv==1.0.1
langchain==0.1.13
langchain-groq==0.1.0
cohere==5.1.0
streamlit==1.32.2
fastapi==0.110.0
uvicorn==0.28.0
httpx==0.27.0
ragas==0.1.7
datasets==2.18.0
